refactor(app): extract session prefetch controller from layout.tsx
Move the session message prefetch queue (LRU eviction, bounded concurrency,
priority warm) out of pages/layout.tsx into
pages/layout/pawwork-session-prefetch.ts as createPawworkSessionPrefetch.

Mechanical extraction: reactive dependencies (params, route, currentDir,
visibleSessionDirs, navigation sessions, global SDK/sync) are passed as a
named dependency object; effects run under the Layout owner unchanged. No
user-visible behavior change.

Refs #1056
